gomath dep as mathlib

+++ examples/sandbox.rs
@@ -5,7 +5,7 @@ struct _Fragment {
     #[fragment_pos]
     pos: Vec4,
     #[flat]
-    texture_id: u8,
+    texture_id: u32,
 }
 
 fn main() {}
